changed insert query to stored procedure

+++ repository/product_interface.go
@@ -5,9 +5,9 @@ import "github.com/shayja/go-template-api/model"
 
 type ProductRepositoryInterface interface {
 	GetAll(page int)([]model.Product, error)
-	GetSingle(uint) (model.Product, error)
-	Create(model.ValidateProduct) (int, error)
-	Update(uint, model.ValidateProduct) (model.Product, error)
-	UpdatePrice(id uint, post model.ValidateProductPrice) (model.Product, error) 
-	Delete(uint) bool
+	GetSingle(int64) (model.Product, error)
+	Create(model.ValidateProduct) (int64, error)
+	Update(int64, model.ValidateProduct) (model.Product, error)
+	UpdatePrice(id int64, post model.ValidateProductPrice) (model.Product, error) 
+	Delete(int64) bool
 }
